Move fixtures in conftest
